requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.3
oauthlib==3.2.0
outcome==1.1.0
Pillow==9.0.1
psycopg-binary==3.0.9
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
schedule==1.1.0
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
wsproto==1.1.0
